to_sql/pyproject.toml
[tool.poetry]
name = "to-sql"
version = "0.1.0"
description = ""
authors = ["Peter Cappetto <peter.cappetto@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pymongo = "^4.6.1"
psycopg2 = "^2.9.9"
python-dotenv = "^1.0.1"
pydantic = "^2.6.1"
typing-extensions = "^4.9.0"
pytest = "^8.0.1"
tqdm = "^4.66.2"
pandas = "^2.2.0"
pyarrow = "^15.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
